+++ app/utils/util.js
@@ -1,6 +1,7 @@
 import {IonicApp, NavController, Alert} from 'ionic-angular';
 import {HTTP_PROVIDERS, Http, Headers, RequestOptions, RequestMethod} from 'angular2/http';
 
+import {SSO} from './sso';
 import {XmlUtil} from './xmlutil';
 import {DateUtil} from './dateutil';
 import {CookieUtil} from './cookieutil';
@@ -56,18 +57,32 @@ export class Util {
     setXMLAttribute(elementNode, attributeNamespace, attributeName, attributeValue) {
         return XmlUtil.setXMLAttribute(elementNode, attributeNamespace, attributeName, attributeValue);
     }
+    
+    callCordysWebserviceUseAnonymous(request) {
+        let useAnonymous = true;
+        return this.callCordysWebservice(request, useAnonymous);
+    }
 
-    callCordysWebservice(request) {
-        return new Promise(resolve => {
+    callCordysWebservice(request, useAnonymous) {
+        if (!useAnonymous) {
+            // If there is not a saml artifact in cookie, then redirect to Login page.
+            let sso = new SSO(this, this.app.config);
+            if (!sso.loggedOn()) {
+                // redirect to Login page.
+                this.app.redirectLoginPage();
+                return;
+            }
+        }
+        return new Promise((resolve, reject) => {
             let url = this.app.config.get("BASE_URL") + this.app.config.get("GATEWAY_URL");
-            if (this.hasCookie(this.app.config.get("SAML_ARTIFACT_COOKIE_NAME"))) {
+            if (!useAnonymous) {
                 url = url + "?" + this.app.config.get("SAMLART_NAME") + "=" +
-                    this.getCookie(this.app.config.get("SAML_ARTIFACT_COOKIE_NAME"));
-
+                this.getCookie(this.app.config.get("SAML_ARTIFACT_COOKIE_NAME"));
                 url = url + "&language=" + this.app.userLang;
             } else {
                 url = url + "?language=" + this.app.userLang;
             }
+            
             this.http.post(url, request)
                 .map(res => res.text())
                 .subscribe(data => {
@@ -90,7 +105,7 @@ export class Util {
                     
                 });
         });
-    }
+    }   
 
     callCordysWebserviceWithUrl(url, request) {
         return new Promise(resolve => {
@@ -101,7 +116,7 @@ export class Util {
                 });
         });
     }
-
+    
     getRequestXml(url) {
         return new Promise(resolve => {
             this.http.get(url)
@@ -128,6 +143,10 @@ export class Util {
 
     hasCookie(name) {
         return CookieUtil.has(name);
+    }
+    
+    transferCordysDateStringToUTC(dateString) {
+        return DateUtil.transferCordysDateStringToUTC(dateString);
     }
 
     getUTCDate() {
